deprecate packPluginFromManifest
