+++ src/cli/plugin/pack.ts
@@ -1,6 +1,6 @@
 import type yargs from "yargs";
 import type { CommandModule } from "yargs";
-import run from "../../plugin/packer/cli";
+import run from "../../plugin/packer/cli/cli";
 import { emitExperimentalWarning } from "../../utils/stability";
 
 const command = "pack";
