+++ src/plugin/packer/from-manifest/__tests__/fixtures/private.ppk
@@ -1,0 +1,15 @@
+-----BEGIN RSA PRIVATE KEY-----
+MIICWwIBAAKBgQDDH5jleaiQagMnSS6XjrOQCJIDhjutOV7fJyzrsGhIGhFnZk7y
+K5YAm+kfOMlCEUZNHavcfS3WrJRP5sanc6E0dobTXNBYVwx97pEZk1KPhPVCiVCC
+HBRPvef+Ai98o4aeSVamP6bD9aIZ51+prjcf0uh9+vzzm1XvqF9rgMkhkQIDAQAB
+AoGAT8GGmaXUxNLQXyqGpORreSBHrrEbi367zLMSpXQB7BbbrkwfUNWIBs+zxlNL
+0HmHJtZN/V4dcnYwWUiXQrL83OHdVgqgcUBKBX5znY9GJcPJoDM6ZhiYR7K79IQq
+eAzKCveFm3bhk9+mnOdiZO+q39IPE+3PPz3dQ1PNRwbCh8ECQQDtKstAHmfRcIwf
+4woOV+TllxRMDkexThxs/YWV+5Sq3lLUqOjJolLFdW1DwqQHjeyeN008C3IhNLyy
+1xeRLLydAkEA0p4gHlCnxUz9nVuRkz2Dk0GMpkdz156Opcf1nKUXar4CbdSp9gbJ
+LEBulEDRVop4nY6J4W5hW0bDjuPGnSR0hQJATpU2YVlxxtjG5S3iQBxpcJVdmVHF
++X7LNmXOZILGoNMnmOUatOy/BkRBXwlYNlSVSVtDkRityUjjGVLhsS2klQJAbkQY
+3qYtX69SK/sPuP2AkCzGPRu1e1JorkSEwzDvlJPL48JuBP9CfWdyPS2+K0etpBdG
++n32fHoM0hdQGV9HnQJAGed+BG6jxfXJe7KRhPoT3bF4lc4xmdXGfLMoutLqoZ8y
+GrKcvq1TcxsA/QSkvSGGEd0667t/9djFq1Ift02RMQ==
+-----END RSA PRIVATE KEY-----
