+++ src/plugin/packer/from-manifest/__tests__/fixtures/sample-plugin/manifest.json
@@ -1,0 +1,9 @@
+{
+  "manifest_version": 1,
+  "version": 1,
+  "type": "APP",
+  "name": {
+    "en": "sample extension"
+  },
+  "icon": "image/icon.png"
+}
