--- src/plugin/packer/pack-plugin-from-manifest.ts
@@ -1,17 +1,0 @@
-import path from "path";
-import packer from "./index";
-import { ManifestFactory } from "./manifest";
-import { ContentsZip } from "./contents-zip";
-import type { PluginZip } from "./plugin-zip";
-
-export const packPluginFromManifest = async (
-  manifestJSONPath: string,
-  privateKey: string,
-): Promise<{ plugin: PluginZip; privateKey: string; id: string }> => {
-  const manifest = await ManifestFactory.loadJsonFile(manifestJSONPath);
-  const contentsZip = await ContentsZip.createFromManifest(
-    path.dirname(manifestJSONPath),
-    manifest,
-  );
-  return packer(contentsZip, privateKey);
-};
